+++ ChangeLog.md
@@ -11,6 +11,7 @@
   ([#96](https://github.com/davep/braindrop/pull/96))
 - Free text search now also looks in the domain of a raindrop.
   ([#96](https://github.com/davep/braindrop/pull/96))
+- Added raindrop types to the navigation panel.
 
 ## v0.5.0